+++ src/components/common/columnActionsPopOverComponents/index.tsx
@@ -2,3 +2,4 @@ export { DateFilterSelect } from './DateFilterSelect';
 export { DateRangeTypeSelect } from './DateRangeTypeSelect';
 export { FilterListSkeleton } from './FilterListSkeleton';
 export { SortSection } from './SortSection';
+export { MultiSelectSection } from './MultiSelectSection';
