+++ src/components/common/paginatedTable/HeaderCell.tsx
@@ -1,4 +1,5 @@
 import { Typography } from '@mui/material';
+
 import { Column } from './PaginatedTable';
 
 interface HeaderCellProps {
