+++ src/hooks/index.ts
@@ -7,5 +7,9 @@ export { useFormUtils } from './useFormUtils';
 export { useNavbarMenuConfig } from './useNavbarMenuConfig';
 export { useBaseLogo } from './useBaseLogo';
 export { useValidationHelper } from './useValidationHelper';
+export { useTableActions } from './useTableActions';
+export { useTableFilters, type TableFilters } from './useTableFilters';
+export { usePaginatedReactTable } from './usePaginatedReactTable';
+export { useMinimalPaginationInfo } from './useMinimalPaginationInfo';
 
 export * from './apiHooks';
